Add files via upload
If that text to audio conversion required document contains a figure, it is not clear to reader-software what the title and figure caption of that figure is. This can lead to information that is being missed or being misinterpreted. Images within webpages have alternative text descriptions, but if this is not available, there is currently no good solution for obtaining text information associated with images within documents.